Make activity log available to hook functions
Adds the activty log to the event model passed to hook functions. This means that hooks can perform actions based on the previous activity recorded on the task.

Use `Object.defineProperty` to assign the activity log to the event model so that the log is non-enumerable by default, and so doesn't get recursively written into the activity log by the activity log hook.

Setting `configurable:true` means that the activity log can be updated as part of a single event to include activty that was created as a result of previous hooks bound to the same event.